+++ ui/src/dashboards/components/DashboardsHeader.tsx
@@ -1,0 +1,18 @@
+import React from 'react'
+
+import SourceIndicator from 'src/shared/components/SourceIndicator'
+
+const DashboardsHeader = (): JSX.Element => (
+  <div className="page-header">
+    <div className="page-header__container">
+      <div className="page-header__left">
+        <h1 className="page-header__title">Dashboards</h1>
+      </div>
+      <div className="page-header__right">
+        <SourceIndicator />
+      </div>
+    </div>
+  </div>
+)
+
+export default DashboardsHeader