+++ ui/src/shared/copy/notifications.ts
@@ -414,6 +414,30 @@ export const notifyDashboardDeleted = name => ({
   message: `Dashboard ${name} deleted successfully.`,
 })
 
+export const notifyDashboardExported = name => ({
+  ...defaultSuccessNotification,
+  icon: 'dash-h',
+  message: `Dashboard ${name} exported successfully.`,
+})
+
+export const notifyDashboardExportFailed = (name, errorMessage) => ({
+  ...defaultErrorNotification,
+  duration: INFINITE,
+  message: `Failed to export Dashboard ${name}: ${errorMessage}.`,
+})
+
+export const notifyDashboardImported = name => ({
+  ...defaultSuccessNotification,
+  icon: 'dash-h',
+  message: `Dashboard ${name} imported successfully.`,
+})
+
+export const notifyDashboardImportFailed = (fileName, errorMessage) => ({
+  ...defaultErrorNotification,
+  duration: INFINITE,
+  message: `Failed to import Dashboard from file ${fileName}: ${errorMessage}.`,
+})
+
 export const notifyDashboardDeleteFailed = (name, errorMessage) =>
   `Failed to delete Dashboard ${name}: ${errorMessage}.`
 
@@ -609,7 +633,7 @@ export const notifyKapacitorNotFound = () => ({
   message: 'We could not find a Kapacitor configuration for this source.',
 })
 
-// IFQL notifications
+// Flux notifications
 export const analyzeSuccess = {
   ...defaultSuccessNotification,
   message: 'No errors found. Happy Happy Joy Joy!',
@@ -621,27 +645,27 @@ export const couldNotGetServices = {
   message: 'We could not get services',
 }
 
-export const ifqlCreated = {
+export const fluxCreated = {
   ...defaultSuccessNotification,
-  message: 'IFQL Connection Created.  Script your heart out!',
+  message: 'Flux Connection Created.  Script your heart out!',
 }
 
-export const ifqlNotCreated = (message: string) => ({
+export const fluxNotCreated = (message: string) => ({
   ...defaultErrorNotification,
   message,
 })
 
-export const ifqlNotUpdated = (message: string) => ({
+export const fluxNotUpdated = (message: string) => ({
   ...defaultErrorNotification,
   message,
 })
 
-export const ifqlUpdated = {
+export const fluxUpdated = {
   ...defaultSuccessNotification,
   message: 'Connection Updated. Rejoice!',
 }
 
-export const ifqlTimeSeriesError = (message: string) => ({
+export const fluxTimeSeriesError = (message: string) => ({
   ...defaultErrorNotification,
   message: `Could not get data: ${message}`,
 })