+++ ui/test/flux/components/From.test.tsx
@@ -1,0 +1,33 @@
+import React from 'react'
+import {shallow} from 'enzyme'
+import From from 'src/flux/components/From'
+import {service} from 'test/resources'
+
+jest.mock('src/shared/apis/metaQuery', () => require('mocks/flux/apis'))
+
+const setup = () => {
+  const props = {
+    funcID: '1',
+    argKey: 'db',
+    value: 'db1',
+    bodyID: '2',
+    declarationID: '1',
+    service,
+    onChangeArg: () => {},
+  }
+
+  const wrapper = shallow(<From {...props} />)
+
+  return {
+    wrapper,
+  }
+}
+
+describe('Flux.Components.From', () => {
+  describe('rendering', () => {
+    it('renders without errors', () => {
+      const {wrapper} = setup()
+      expect(wrapper.exists()).toBe(true)
+    })
+  })
+})